% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwrpv.R
\name{highlow}
\alias{highlow}
\title{highlow()}
\usage{
highlow(downtrim, n, pa, pb)
}
\arguments{
\item{downtrim}{lower bound}

\item{n}{upper bound}

\item{pa}{parameter of the mixture distribution, a real number between zero and one with pa+pb=1}

\item{pb}{parameter of the mixture distribution, a real number between zero and one with pa+pb=1}
}
\value{
c(lhigh, llow))  # return the new upper and lower bounds
}
\description{
If possible, trim the upper and lower bounds
}

