% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwrpv.R
\name{regresults}
\alias{regresults}
\title{regresults: sample data}
\format{csv format file with 4 variables (beta, n0, n1, n2) and 120 rows}
\description{
A sample dataset of input regression results based on machine-level accurate cumulative normal values.
Rather than just typing in a few digits of the 2.5%, .5% and 2.5%x10^(-6) normal p-value cutoffs,
the norminverse function in RATS was used to create sample-case betas which are exact
}

